+++ changelog.md
@@ -26,11 +26,12 @@
 * **breaking change** The API endpoint `/api/database/formula/<field_id>/type/` now requires
   `table_id` instead of `field_id`, and also `name` in the request body.
 * Added support in dev.sh for KDE's Konsole terminal emulator.
-* Fixed a bug that truncated characters for email in the sidebar.
 * Fixed a bug that would sometimes cancel multi-cell selection.
 * Upgraded node runtime to v16.14.0
 * Cache aggregation values to improve performances
 * Added new endpoint to get all configured aggregations for a grid view
+* Fixed DONT_UPDATE_FORMULAS_AFTER_MIGRATION env var not working correctly.
+* Stopped the generated model cache clear operation also deleting all other redis keys.
 
 ## Released (2022-03-03 1.9.1)